requirements.txt
matplotlib==3.0.2
numpy==1.16.2
numpy-stl==2.10.0
tqdm==4.31.1
